+++ Include/internal/slp_platformselect.h
@@ -6,31 +6,31 @@
  */
 
 #if   defined(MS_WIN32) && !defined(MS_WIN64) && defined(_M_IX86)
-#include "switch_x86_msvc.h" /* MS Visual Studio on X86 */
+#include "internal/slp_switch_x86_msvc.h" /* MS Visual Studio on X86 */
 #elif defined(MS_WIN64) && defined(_M_X64)
-#include "switch_x64_msvc.h" /* MS Visual Studio on X64 */
+#include "internal/slp_switch_x64_msvc.h" /* MS Visual Studio on X64 */
 #elif defined(__GNUC__) && defined(__i386__)
-#include "switch_x86_unix.h" /* gcc on X86 */
+#include "internal/slp_switch_x86_unix.h" /* gcc on X86 */
 #elif defined(__GNUC__) && defined(__amd64__)
-#include "switch_amd64_unix.h" /* gcc on amd64 */
+#include "internal/slp_switch_amd64_unix.h" /* gcc on amd64 */
 #elif defined(__GNUC__) && defined(__PPC__) && defined(__linux__)
-#include "switch_ppc_unix.h" /* gcc on PowerPC */
+#include "internal/slp_switch_ppc_unix.h" /* gcc on PowerPC */
 #elif defined(__GNUC__) && defined(__ppc__) && defined(__APPLE__)
-#include "switch_ppc_macosx.h" /* Apple MacOS X on PowerPC */
+#include "internal/slp_switch_ppc_macosx.h" /* Apple MacOS X on PowerPC */
 #elif defined(__GNUC__) && defined(sparc) && defined(sun)
-#include "switch_sparc_sun_gcc.h" /* SunOS sparc with gcc */
+#include "internal/slp_switch_sparc_sun_gcc.h" /* SunOS sparc with gcc */
 #elif defined(__GNUC__) && defined(__s390__) && defined(__linux__)
-#include "switch_s390_unix.h"   /* Linux/S390 */
+#include "internal/slp_switch_s390_unix.h"   /* Linux/S390 */
 #elif defined(__GNUC__) && defined(__s390x__) && defined(__linux__)
-#include "switch_s390_unix.h"   /* Linux/S390 zSeries (identical) */
+#include "internal/slp_switch_s390_unix.h"   /* Linux/S390 zSeries (identical) */
 #elif defined(__GNUC__) && defined(__arm__) && defined(__thumb__)
-#include "switch_arm_thumb_gcc.h" /* gcc using arm thumb */
+#include "internal/slp_switch_arm_thumb_gcc.h" /* gcc using arm thumb */
 #elif defined(__GNUC__) && defined(__arm32__)
-#include "switch_arm32_gcc.h" /* gcc using arm32 */
+#include "internal/slp_switch_arm32_gcc.h" /* gcc using arm32 */
 #elif defined(__GNUC__) && defined(__mips__) && defined(__linux__)
-#include "switch_mips_unix.h" /* MIPS */
+#include "internal/slp_switch_mips_unix.h" /* MIPS */
 #elif defined(SN_TARGET_PS3)
-#include "switch_ps3_SNTools.h" /* Sony PS3 */
+#include "internal/slp_switch_ps3_SNTools.h" /* Sony PS3 */
 #endif
 
 /* default definitions if not defined in above files */
