+++ Objects/descrobject.c
@@ -3,7 +3,7 @@
 #include "Python.h"
 #include "internal/pystate.h"
 #include "structmember.h" /* Why is this not included in Python.h? */
-#include "core/stackless_impl.h"
+#include "internal/stackless_impl.h"
 
 /*[clinic input]
 class mappingproxy "mappingproxyobject *" "&PyDictProxy_Type"
