+++ PCbuild/pythoncore.vcxproj
@@ -213,20 +213,18 @@
     <ClInclude Include="..\Python\importdl.h" />
     <ClInclude Include="..\Python\thread_nt.h" />
     <ClInclude Include="..\Python\wordcode_helpers.h" />
-    <ClInclude Include="..\Stackless\core\cframeobject.h" />
-    <ClInclude Include="..\Stackless\core\slp_exttype.h" />
-    <ClInclude Include="..\Stackless\core\stackless_impl.h" />
-    <ClInclude Include="..\Stackless\core\stackless_structs.h" />
-    <ClInclude Include="..\Stackless\core\stackless_tstate.h" />
-    <ClInclude Include="..\Stackless\core\stackless_pystate.h" />
-    <ClInclude Include="..\Stackless\module\channelobject.h" />
-    <ClInclude Include="..\Stackless\pickling\prickelpit.h" />
-    <ClInclude Include="..\Stackless\platf\slp_platformselect.h" />
-    <ClInclude Include="..\Stackless\platf\switch_x64_msvc.h" />
-    <ClInclude Include="..\Stackless\platf\switch_x86_msvc.h" />
-    <ClInclude Include="..\Stackless\stackless.h" />
-    <ClInclude Include="..\Stackless\stackless_api.h" />
-    <ClInclude Include="..\Stackless\stackless_version.h" />
+    <ClInclude Include="..\Include\slp_exttype.h" />
+    <ClInclude Include="..\Include\internal\stackless_impl.h" />
+    <ClInclude Include="..\Include\stackless_structs.h" />
+    <ClInclude Include="..\Include\stackless_tstate.h" />
+    <ClInclude Include="..\Include\internal\stackless_pystate.h" />
+    <ClInclude Include="..\Include\internal\slp_prickelpit.h" />
+    <ClInclude Include="..\Include\internal\slp_platformselect.h" />
+    <ClInclude Include="..\Include\internal\slp_switch_x64_msvc.h" />
+    <ClInclude Include="..\Include\internal\slp_switch_x86_msvc.h" />
+    <ClInclude Include="..\Include\stackless.h" />
+    <ClInclude Include="..\Include\stackless_api.h" />
+    <ClInclude Include="..\Include\stackless_version.h" />
   </ItemGroup>
   <ItemGroup Condition="$(IncludeExternals)">
     <ClInclude Include="$(zlibDir)\crc32.h" />
