--- Stackless/core/cframeobject.h
@@ -1,1 +1,0 @@
-/* header is now empty and can be removed. */
