+++ Stackless/core/stacklesseval.c
@@ -5,11 +5,11 @@
 #include "frameobject.h"
 #include "structmember.h"
 
-#include "stackless_impl.h"
-#include "pickling/prickelpit.h"
+#include "internal/stackless_impl.h"
+#include "internal/slp_prickelpit.h"
 
 /* platform specific constants */
-#include "platf/slp_platformselect.h"
+#include "internal/slp_platformselect.h"
 
 /* Stackless extension for ceval.c */
