+++ Stackless/module/scheduling.c
@@ -2,7 +2,7 @@
 #include "structmember.h"
 
 #ifdef STACKLESS
-#include "core/stackless_impl.h"
+#include "internal/stackless_impl.h"
 
 #include "pythread.h"
