+++ Stackless/pickling/prickelpit.c
@@ -4,12 +4,11 @@
 #include <stddef.h>  /* for offsetof() */
 #include "compile.h"
 
-#include "core/stackless_impl.h"
-#include "pickling/prickelpit.h"
-#include "module/channelobject.h"
+#include "internal/stackless_impl.h"
+#include "internal/slp_prickelpit.h"
 
 /* platform specific constants */
-#include "platf/slp_platformselect.h"
+#include "internal/slp_platformselect.h"
 
 /******************************************************
